Scraping now works (barely)

+++ gdg/scrape.py
@@ -5,45 +5,59 @@ from peewee import *
 import gdg
 from gdg.data import *
 
-config = ConfigParser()
+config = ConfigParser.ConfigParser()
 
 def get_files(path):
+    images = []
     for root, dirs, files in os.walk(path):
         for filename in files:
-            yield os.path.join(root, filename)
+            images.append(os.path.join(root, filename))
+    return images
 
 def get_thumb(path, base, ext):
-    return os.path.join(path, base + ext);
+    return os.path.join(path, base + ext)
 
-def main():
+def get_directory(path):
+    path = path.translate(None, '"\'')
+    if (path == None or path == 'gdg.current_dir'):
+        return gdg.current_dir
+    if (os.path.isdir(path)):
+        return os.path.abspath(path)
+    return os.path.join(gdg.current_dir, path)
+
+def scrape_images():
     config.read(os.path.join(gdg.current_dir, 'gdg.conf'))
-    dbpath = config['database']['path']
+    dbpath = get_directory(config.get('database', 'path'))
+
+    print(dbpath)
     
     with GoddamnDatabase(dbpath) as db:
-        current = Image.select(Image.path)
-        ondisk = get_files(config['images']['path'])
-        new = set(ondisk) - set(current)
-        deleted = set(current) - set(ondisk)
+        current = [i.path for i in Image.select()]
+        ondisk = get_files(get_directory(config.get('images', 'path')))
+        new_files = set(ondisk) - set(current)
+        deleted_files = set(current) - set(ondisk)
         
-        with db.transaction():
-            for f in new:
-                i = Image()
-                i.path = f
-                i.save()
-            
-        with db.transaction():
-            for f in deleted:
-                Image.delete().where(Image.path == f).execute()
+        if len(new_files) > 0:
+            with db.transaction():
+                for f in new_files:
+                    i = Image()
+                    i.path = f
+                    i.save()
+                    
+        if len(deleted_files) > 0:
+            with db.transaction():
+                for f in deleted_files:
+                    Image.delete().where(Image.path == f).execute()
     
     with GoddamnDatabase(dbpath) as db:
         for i in Image.select().where(Image.thumb == None):
-            img = Image.open(PIL.Image.open(i.path))
+            img = PIL.Image.open(i.path)
             x = 0
             w = min(img.size)
             h = w
             
             if (img.size[0] > img.size[1]):
-                x = int((i.size[0] - w) / 2)
+                x = int((img.size[0] - w) / 2)
             
             box = (x, 0, x + w, h)
             
@@ -52,15 +66,16 @@ def main():
             
             path_parts = os.path.split(i.path)
             name_parts = os.path.splitext(path_parts[1])
+            thumb_path = get_directory(config.get('thumbnails', 'path'))
             
-            thumb_name = config['thumbnails']['prefix'] + name_parts[0] + config['thumbnails']['postfix']
+            thumb_name = config.get('thumbnails', 'prefix').translate(None, '"\'') + name_parts[0] + config.get('thumbnails', 'postfix').translate(None, '"\'')
             
-            thumb = get_thumb(path_parts[0], thumb_name, name_parts[1])
+            thumb = get_thumb(thumb_path, thumb_name, name_parts[1])
             
             counter = 1
             while os.path.isfile(thumb):
                 thumb_name = thumb_name + counter
-                thumb = get_thumb(path_parts[0], thumb_name, name_parts[1])
+                thumb = get_thumb(thumb_path, thumb_name, name_parts[1])
                 
             t.save(thumb)
